+++ archives-manage-dao/src/main/java/com/yintu/rixing/archives/ArchArchivesQuantityStatisticsMapper.java
@@ -2,6 +2,7 @@ package com.yintu.rixing.archives;
 
 import org.apache.ibatis.annotations.Mapper;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
@@ -24,5 +25,5 @@ public interface ArchArchivesQuantityStatisticsMapper {
      * @param endDate           结束日期
      * @return ..
      */
-    Map<String, Long> selectArchivesQuantityStatisticsData(Integer archivesLibraryId, String tableName, Integer departmentId, Date startDate, Date endDate);
+    Map<String, BigDecimal> selectArchivesQuantityStatisticsData(Integer archivesLibraryId, String tableName, Integer departmentId, Date startDate, Date endDate);
 }
